[SPARK-13019] move new files into mllib folder
