+++ examples/src/main/java/org/apache/spark/examples/mllib/JavaHypothesisTestingKolmogorovSmirnovTestExample.java
@@ -15,14 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.spark.examples.ml;
+package org.apache.spark.examples.mllib;
 
 // $example on$
-import org.apache.spark.api.java.JavaRDD;
+import java.util.Arrays;
+
+import org.apache.spark.api.java.JavaDoubleRDD;
 import org.apache.spark.api.java.JavaSparkContext;
-import org.apache.spark.mllib.linalg.Vector;
-import org.apache.spark.mllib.stat.MultivariateStatisticalSummary;
+
 import org.apache.spark.mllib.stat.Statistics;
+import org.apache.spark.mllib.stat.test.KolmogorovSmirnovTestResult;
 // $example off$
 
 import org.apache.spark.SparkConf;
@@ -31,27 +33,27 @@ import org.apache.spark.mllib.linalg.Vectors;
 import java.util.Arrays;
 
 
-public class JavaSummaryStatisticsExample {
+public class JavaHypothesisTestingKolmogorovSmirnovTestExample {
     public static void main(String[] args) {
 
-        SparkConf conf = new SparkConf().setAppName("JavaSummaryStatisticsExample");
+        SparkConf conf = new SparkConf().setAppName("JavaHypothesisTestingKolmogorovSmirnovTestExample");
         JavaSparkContext jsc = new JavaSparkContext(conf);
         SQLContext sqlContext = new SQLContext(jsc);
 
         // $example on$
-        Vector v1 = Vectors.dense(1.0, 10.0, 100.0);
-        Vector v2 = Vectors.dense(2.0, 20.0, 200.0);
-        Vector v3 = Vectors.dense(3.0, 30.0, 300.0);
 
-        JavaRDD<Vector> mat = jsc.parallelize(Arrays.asList(v1, v2, v3)); // an RDD of Vectors
+        // @note: todo
 
-        // Compute column summary statistics.
-        MultivariateStatisticalSummary summary = Statistics.colStats(mat.rdd());
-        System.out.println(summary.mean()); // a dense vector containing the mean value for each column
-        System.out.println(summary.variance()); // column-wise variance
-        System.out.println(summary.numNonzeros()); // number of nonzeros in each column
+        JavaDoubleRDD data = jsc.parallelizeDoubles(Arrays.asList(0.2, 1.0, ...));
+        KolmogorovSmirnovTestResult testResult = Statistics.kolmogorovSmirnovTest(data, "norm", 0.0, 1.0);
+        // summary of the test including the p-value, test statistic,
+        // and null hypothesis
+        // if our p-value indicates significance, we can reject the null hypothesis
+        System.out.println(testResult);
+
         // $example off$
 
         jsc.stop();
     }
 }
+
