+++ examples/src/main/java/org/apache/spark/examples/mllib/JavaSummaryStatisticsExample.java
@@ -15,13 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.spark.examples.ml;
+package org.apache.spark.examples.mllib;
 
 // $example on$
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
-import org.apache.spark.mllib.stat.KernelDensity;
-import org.apache.spark.rdd.RDD;
+import org.apache.spark.mllib.linalg.Vector;
+import org.apache.spark.mllib.stat.MultivariateStatisticalSummary;
+import org.apache.spark.mllib.stat.Statistics;
 // $example off$
 
 import org.apache.spark.SparkConf;
@@ -30,30 +31,27 @@ import org.apache.spark.mllib.linalg.Vectors;
 import java.util.Arrays;
 
 
-public class JavaKernelDensityEstimationExample {
+public class JavaSummaryStatisticsExample {
     public static void main(String[] args) {
 
-        SparkConf conf = new SparkConf().setAppName("JavaKernelDensityEstimationExample");
+        SparkConf conf = new SparkConf().setAppName("JavaSummaryStatisticsExample");
         JavaSparkContext jsc = new JavaSparkContext(conf);
         SQLContext sqlContext = new SQLContext(jsc);
 
         // $example on$
+        Vector v1 = Vectors.dense(1.0, 10.0, 100.0);
+        Vector v2 = Vectors.dense(2.0, 20.0, 200.0);
+        Vector v3 = Vectors.dense(3.0, 30.0, 300.0);
 
-        // @note: todo
+        JavaRDD<Vector> mat = jsc.parallelize(Arrays.asList(v1, v2, v3)); // an RDD of Vectors
 
-        RDD<Double> data = ... // an RDD of sample data
-
-        // Construct the density estimator with the sample data and a standard deviation for the Gaussian
-        // kernels
-        KernelDensity kd = new KernelDensity()
-                .setSample(data)
-                .setBandwidth(3.0);
-
-        // Find density estimates for the given values
-        double[] densities = kd.estimate(new double[] {-1.0, 2.0, 5.0});
+        // Compute column summary statistics.
+        MultivariateStatisticalSummary summary = Statistics.colStats(mat.rdd());
+        System.out.println(summary.mean()); // a dense vector containing the mean value for each column
+        System.out.println(summary.variance()); // column-wise variance
+        System.out.println(summary.numNonzeros()); // number of nonzeros in each column
         // $example off$
 
         jsc.stop();
     }
 }
-
