+++ examples/src/main/python/mllib/stratified_sampling_example.py
