--- examples/src/main/scala/org/apache/spark/examples/ml/SummaryStatisticsExample.scala
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// scalastyle:off println
-package org.apache.spark.examples.ml
-
-// $example on$
-import org.apache.spark.mllib.linalg.Vectors
-import org.apache.spark.mllib.stat.{MultivariateStatisticalSummary, Statistics}
-// $example off$
-import org.apache.spark.sql.SQLContext
-import org.apache.spark.{SparkConf, SparkContext}
-
-object SummaryStatisticsExample {
-
-  def main(args: Array[String]) {
-
-    val conf = new SparkConf().setAppName("SummaryStatisticsExample").setMaster("local[*]")
-    val sc = new SparkContext(conf)
-    val sqlContext = new SQLContext(sc)
-
-    // $example on$
-    val v1 = Vectors.dense(1.0, 10.0, 100.0)
-    val v2 = Vectors.dense(2.0, 20.0, 200.0)
-    val v3 = Vectors.dense(3.0, 30.0, 300.0)
-
-    val observations = sc.parallelize(Seq(v1, v2, v3))
-
-    // Compute column summary statistics.
-    val summary: MultivariateStatisticalSummary = Statistics.colStats(observations)
-    println(summary.mean) // a dense vector containing the mean value for each column
-    println(summary.variance) // column-wise variance
-    println(summary.numNonzeros) // number of nonzeros in each column
-    // $example off$
-
-    sc.stop()
-  }
-}
-// scalastyle:on println
