+++ examples/src/main/scala/org/apache/spark/examples/mllib/HypothesisTestingKolmogorovSmirnovTestExample.scala
@@ -16,39 +16,39 @@
  */
 
 // scalastyle:off println
-package org.apache.spark.examples.ml
+package org.apache.spark.examples.mllib
 
 // $example on$
-import org.apache.spark.mllib.linalg._
 import org.apache.spark.mllib.stat.Statistics
-// $example off$
 import org.apache.spark.rdd.RDD
+
+// $example off$
 import org.apache.spark.sql.SQLContext
 import org.apache.spark.{SparkConf, SparkContext}
 
-object CorrelationsExample {
+object HypothesisTestingKolmogorovSmirnovTestExample {
 
   def main(args: Array[String]) {
 
-    val conf = new SparkConf().setAppName("CorrelationsExample").setMaster("local[*]")
+    val conf = new SparkConf().setAppName("HypothesisTestingKolmogorovSmirnovTestExample").setMaster("local[*]")
     val sc = new SparkContext(conf)
     val sqlContext = new SQLContext(sc)
 
     // $example on$
-    val seriesX: RDD[Double] = sc.parallelize(Array(1, 2, 3, 3, 5)) // a series
-    val seriesY: RDD[Double] = sc.parallelize(Array(11, 22, 33, 33, 555)) // must have the same number of partitions and cardinality as seriesX
+    // @note: todo
 
-    // compute the correlation using Pearson's method. Enter "spearman" for Spearman's method. If a
-    // method is not specified, Pearson's method will be used by default.
-    val correlation: Double = Statistics.corr(seriesX, seriesY, "pearson")
-    println(correlation)
+    val data: RDD[Double] = ... // an RDD of sample data
 
-    val data: RDD[Vector] = sc.parallelize(Seq(Vectors.dense(1.0, 10.0, 100.0), Vectors.dense(2.0, 20.0, 200.0), Vectors.dense(5.0, 33.0, 366.0))) // note that each Vector is a row and not a column
+    // run a KS test for the sample versus a standard normal distribution
+    val testResult = Statistics.kolmogorovSmirnovTest(data, "norm", 0, 1)
+    println(testResult) // summary of the test including the p-value, test statistic,
+    // and null hypothesis
+    // if our p-value indicates significance, we can reject the null hypothesis
 
-    // calculate the correlation matrix using Pearson's method. Use "spearman" for Spearman's method.
-    // If a method is not specified, Pearson's method will be used by default.
-    val correlMatrix: Matrix = Statistics.corr(data, "pearson")
-    println(correlMatrix.toString)
+    // perform a KS test using a cumulative distribution function of our making
+    val myCDF: Double => Double = ...
+    val testResult2 = Statistics.kolmogorovSmirnovTest(data, myCDF)
+
     // $example off$
 
     sc.stop()
