+++ examples/src/main/scala/org/apache/spark/examples/mllib/KernelDensityEstimationExample.scala
@@ -16,20 +16,20 @@
  */
 
 // scalastyle:off println
-package org.apache.spark.examples.ml
+package org.apache.spark.examples.mllib
 
 // $example on$
-import org.apache.spark.mllib.random.RandomRDDs._
-
+import org.apache.spark.mllib.stat.KernelDensity
+import org.apache.spark.rdd.RDD
 // $example off$
 import org.apache.spark.sql.SQLContext
 import org.apache.spark.{SparkConf, SparkContext}
 
-object RandomDataGenerationExample {
+object KernelDensityEstimationExample {
 
   def main(args: Array[String]) {
 
-    val conf = new SparkConf().setAppName("RandomDataGenerationExample").setMaster("local[*]")
+    val conf = new SparkConf().setAppName("KernelDensityEstimationExample").setMaster("local[*]")
     val sc = new SparkContext(conf)
     val sqlContext = new SQLContext(sc)
 
@@ -37,12 +37,16 @@ object RandomDataGenerationExample {
 
     // @note: todo
 
-    // Generate a random double RDD that contains 1 million i.i.d. values drawn from the
-    // standard normal distribution `N(0, 1)`, evenly distributed in 10 partitions.
-    val u = normalRDD(sc, 1000000L, 10)
-    // Apply a transform to get a random double RDD following `N(1, 4)`.
-    val v = u.map(x => 1.0 + 2.0 * x)
+    val data: RDD[Double] = ... // an RDD of sample data
 
+    // Construct the density estimator with the sample data and a standard deviation for the Gaussian
+    // kernels
+    val kd = new KernelDensity()
+      .setSample(data)
+      .setBandwidth(3.0)
+
+    // Find density estimates for the given values
+    val densities = kd.estimate(Array(-1.0, 2.0, 5.0))
     // $example off$
 
     sc.stop()
